# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Superhéroes y Villanos",
    "version": "1.0.0",
    "description": "API para gestionar superhéroes, villanos y enfrentamientos."
  },
  "servers": [
    { "url": "http://localhost:3001" }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Ingresa tu token JWT como: Bearer tu_token_aqui"
      }
    }
  },
  "security": [
    { "Bearer": [] }
  ],
  "paths": {
    "/api/register": {
      "post": {
        "summary": "Registrar usuario",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario registrado" }
        }
      }
    },
    "/api/login": {
      "post": {
        "summary": "Iniciar sesión",
        "tags": ["Auth"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token JWT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/heroes": {
      "get": {
        "summary": "Obtener todos los héroes",
        "tags": ["Héroes"],
        "responses": {
          "200": {
            "description": "Lista de héroes",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Hero" } }
              }
            }
          },
          "401": {
            "description": "Token requerido"
          }
        }
      },
      "post": {
        "summary": "Crear un héroe",
        "tags": ["Héroes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Hero" }
            }
          }
        },
        "responses": {
          "201": { "description": "Héroe creado" },
          "401": {
            "description": "Token requerido"
          }
        }
      }
    },
    "/api/heroes/{id}": {
      "put": {
        "summary": "Actualizar un héroe",
        "tags": ["Héroes"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Hero" }
            }
          }
        },
        "responses": { 
          "200": { "description": "Héroe actualizado" },
          "401": {
            "description": "Token requerido"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un héroe",
        "tags": ["Héroes"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { 
          "200": { "description": "Héroe eliminado" },
          "401": {
            "description": "Token requerido"
          }
        }
      }
    },
    "/api/villains": {
      "get": {
        "summary": "Obtener todos los villanos",
        "tags": ["Villanos"],
        "responses": {
          "200": {
            "description": "Lista de villanos",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Villain" } }
              }
            }
          },
          "401": {
            "description": "Token requerido"
          }
        }
      },
      "post": {
        "summary": "Crear un villano",
        "tags": ["Villanos"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Villain" }
            }
          }
        },
        "responses": {
          "201": { "description": "Villano creado" },
          "401": {
            "description": "Token requerido"
          }
        }
      }
    },
    "/api/villains/{id}": {
      "put": {
        "summary": "Actualizar un villano",
        "tags": ["Villanos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Villain" }
            }
          }
        },
        "responses": { 
          "200": { "description": "Villano actualizado" },
          "401": {
            "description": "Token requerido"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un villano",
        "tags": ["Villanos"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { 
          "200": { "description": "Villano eliminado" },
          "401": {
            "description": "Token requerido"
          }
        }
      }
    },
    "/api/battles": {
      "get": {
        "summary": "Obtener todos los enfrentamientos",
        "tags": ["Batallas"],
        "responses": {
          "200": {
            "description": "Lista de enfrentamientos",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Battle" } }
              }
            }
          },
          "401": {
            "description": "Token requerido"
          }
        }
      },
      "post": {
        "summary": "Crear un enfrentamiento",
        "tags": ["Batallas"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "heroId": { "type": "string" },
                  "villainId": { "type": "string" },
                  "winner": { 
                    "type": "string",
                    "enum": ["hero", "villain"]
                  }
                },
                "required": ["heroId", "villainId", "winner"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Enfrentamiento creado" },
          "401": {
            "description": "Token requerido"
          }
        }
      }
    },
    "/api/battles/turn-based-teams": {
      "post": {
        "summary": "Crear batalla 3vs3 manual",
        "description": "Crea una batalla 3vs3 entre equipos de héroes y villanos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "heroTeam": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "characterId": { "type": "integer", "example": 1 },
                        "type": { "type": "string", "example": "hero" }
                      },
                      "required": ["characterId", "type"]
                    }
                  },
                  "villainTeam": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "characterId": { "type": "integer", "example": 1 },
                        "type": { "type": "string", "example": "villain" }
                      },
                      "required": ["characterId", "type"]
                    }
                  }
                },
                "required": ["heroTeam", "villainTeam"],
                "example": {
                  "heroTeam": [
                    {"characterId": 1, "type": "hero"},
                    {"characterId": 2, "type": "hero"},
                    {"characterId": 3, "type": "hero"}
                  ],
                  "villainTeam": [
                    {"characterId": 1, "type": "villain"},
                    {"characterId": 2, "type": "villain"},
                    {"characterId": 3, "type": "villain"}
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Batalla creada",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TurnBasedTeamBattle" }
              }
            }
          },
          "400": { "description": "Datos inválidos" }
        }
      }
    },
    "/api/battles/turn-based-teams/{battleId}": {
      "get": {
        "summary": "Obtener batalla 3vs3 manual por ID",
        "parameters": [
          {
            "name": "battleId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Estado de la batalla",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TurnBasedTeamBattle" }
              }
            }
          },
          "404": { "description": "Batalla no encontrada" }
        }
      }
    },
    "/api/battles/turn-based-teams/{battleId}/attack": {
      "post": {
        "summary": "Realizar ataque manual en batalla 3vs3",
        "description": "Permite que un personaje de un equipo ataque a un personaje del equipo contrario. El ataque puede ser básico o especial.",
        "parameters": [
          {
            "name": "battleId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "attackerType": { "type": "string", "example": "hero" },
                  "attackType": { "type": "string", "example": "basic" }
                },
                "required": ["attackerType", "attackType"],
                "example": {
                  "attackerType": "hero",
                  "attackType": "basic"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado del ataque",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TurnBasedTeamBattle" }
              }
            }
          },
          "400": { "description": "Datos inválidos" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Hero": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "alias": { "type": "string" },
          "city": { "type": "string" },
          "team": { "type": "string" }
        },
        "required": ["name", "alias"]
      },
      "Villain": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "alias": { "type": "string" },
          "city": { "type": "string" },
          "team": { "type": "string" }
        },
        "required": ["name", "alias"]
      },
      "Battle": {
        "type": "object",
        "properties": {
          "heroId": { "type": "string" },
          "villainId": { "type": "string" },
          "winner": { 
            "type": "string",
            "enum": ["hero", "villain"]
          }
        },
        "required": ["heroId", "villainId", "winner"]
      },
      "TurnBasedTeamBattle": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "type": { "type": "string", "example": "turn-based-teams-manual" },
          "heroTeam": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" },
                "life": { "type": "integer" },
                "maxLife": { "type": "integer" },
                "shield": { "type": "integer" },
                "maxShield": { "type": "integer" },
                "powerBar": { "type": "integer" }
              }
            }
          },
          "villainTeam": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "name": { "type": "string" },
                "life": { "type": "integer" },
                "maxLife": { "type": "integer" },
                "shield": { "type": "integer" },
                "maxShield": { "type": "integer" },
                "powerBar": { "type": "integer" }
              }
            }
          },
          "turns": {
            "type": "array",
            "items": { "type": "object" }
          },
          "status": { "type": "string" },
          "winner": { "type": ["string", "null"] },
          "date": { "type": "string", "format": "date-time" }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Ingresa tu token JWT como: Bearer tu_token_aqui"
      }
    }
  }
} 